foto banner handler, and recovering password mailer

+++ src/main/java/clicker/back/controllers/utils/FotoBannerController.java
@@ -1,12 +1,15 @@
 package clicker.back.controllers.utils;
 
 import clicker.back.Setup;
+import clicker.back.services.AmazonService;
 import clicker.back.utils.entities.FotoBanner;
+import clicker.back.utils.errors.ResponseService;
 import clicker.back.utils.services.FotoBannerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -16,6 +19,9 @@ import java.util.List;
 public class FotoBannerController {
     @Autowired
     FotoBannerService fotoBannerService;
+
+    @Autowired
+    AmazonService amazonService;
 
     @GetMapping
     @ResponseBody
@@ -29,12 +35,18 @@ public class FotoBannerController {
 
     @PostMapping
     @ResponseBody
-    public ResponseEntity<Object> postFoto(@RequestBody FotoBanner fotoBanner){
-        List<FotoBanner> foto = fotoBannerService.findAll();
-        for (FotoBanner banner : foto) {
-            fotoBannerService.delete(banner);
-        }
+    public ResponseEntity<Object> postFoto(@RequestPart("file")MultipartFile multipartFile){
         try{
+            if(multipartFile==null){
+                return ResponseService.genError("no se envio imagen",HttpStatus.BAD_REQUEST);
+            }
+            List<FotoBanner> foto = fotoBannerService.findAll();
+            for (FotoBanner banner : foto) {
+                amazonService.deleteFileFromS3Bucket(banner.getFoto());
+                fotoBannerService.delete(banner);
+            }
+            FotoBanner fotoBanner=new FotoBanner();
+            fotoBanner.setFoto(amazonService.uploadFile(multipartFile,"server","fotoBanner"));
             return new ResponseEntity<>(fotoBannerService.save(fotoBanner),HttpStatus.OK);
         }catch (Exception e){
             return new ResponseEntity<>("FALLO",HttpStatus.BAD_REQUEST);
